+++ src/components/Button/docs/demos/TextAndLink.vue
@@ -7,6 +7,7 @@ import { MButton } from 'morya-ui'
     <MButton label="Text" text />
     <MButton label="Text Danger" text severity="danger" />
     <MButton label="Link" link />
+    <MButton label="Link No Underline" link :underline="false" />
     <MButton label="Link Secondary" link severity="secondary" />
   </div>
 </template>